backend/requirements.txt
asgiref==3.11.0
Django==6.0.1
django-ckeditor==6.7.3
django-cors-headers==4.9.0
django-filter==24.2
django-js-asset==3.1.2
djangorestframework==3.16.1
pillow==12.1.0
psycopg2-binary==2.9.11
sqlparse==0.5.5
tzdata==2025.3
dj-database-url
gunicorn
whitenoise
